api-gateway/requirements.txt
"""
Requirements for API Gateway
"""

# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Celery (for task dispatch)
celery==5.3.4

# HTTP client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
